+++ src/lib.rs
@@ -10,15 +10,54 @@
 //!
 //! Unit types can easily be converted to and from different denominations.
 //!
+//! Ohm's Law is implemented for `Voltage` and `Current` types, allowing you to easily calculate
+//! between the three units using the `/` and `*` operators.
+//!
+//! ## Examples
+//!
+//! Determine the resistance of a 5V, 1A load:
+//! ```rust
+//! use ohms::prelude::*;
+//!
+//! let voltage = 5.volts();
+//! let current = 1u32.amps();
+//!
+//! let resistance = voltage / current; // 5V / 1A = 5Ω
+//! println!("Resistance: {} Ω", resistance.ohms());
+//! ```
+//!
+//! Determine the current of a 5V, 220Ω load:
+//! ```rust
+//! use ohms::prelude::*;
+//!
+//! let voltage = 5.volts();
+//! let resistance = 220.ohms();
+//!
+//! let current = voltage / resistance; // 5V / 220Ω = 22.72mA
+//! println!("Current: {} mA", current.milli_amps());
+//! ```
+//!
+//! Determine the voltage of a 0.5A, 4.7kΩ load:
+//! ```rust
+//! use ohms::prelude::*;
+//!
+//! let current = 0.5f32.milli_amps();
+//! let resistance = 4.7f32.kilo_ohms();
+//!
+//! let voltage = current * resistance; // 0.5mA * 4.7kΩ = 2.35V
+//! println!("Voltage: {} V", voltage.volts());
+//! ```
 #![no_std]
 
 mod assert;
 mod current;
 mod helpers;
+mod law;
 pub mod prelude;
 mod resistance;
 mod voltage;
 
 pub use current::{Current, ExtF32 as CurrentExtF32, ExtU32 as CurrentExtU32};
+pub use law::*;
 pub use resistance::{ExtF32 as ResistanceExtF32, ExtU32 as ResistanceExtU32, Resistance};
 pub use voltage::{ExtF32 as VoltageExtF32, ExtI32 as VoltageExtI32, Voltage};