+++ src/prelude.rs
@@ -1,3 +1,4 @@
 pub use crate::current::{Current, ExtF32 as CurrentExtF32, ExtU32 as CurrentExtU32};
+pub use crate::law::*;
 pub use crate::resistance::{ExtF32 as ResistanceExtF32, ExtU32 as ResistanceExtU32, Resistance};
 pub use crate::voltage::{ExtF32 as VoltageExtF32, ExtI32 as VoltageExtI32, Voltage};