+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-02-17
+
+### Added
+
+- Ohm's Law calculations for `Voltage`, `Current`, and `Resistance` via operators
+- Unit tests for Ohm's Law calculations
+
 ## [0.1.4] - 2023-02-16
 
 ### Added